Fix bug with period crossing months
Seems like a period is more about specifying units of measurement,
e.g. '1 month, 3 weeks and 5 seconds', but when trying to convert
that measurement into seconds, it borks - it doesn't know how
many seconds are in a month. That's reasonable.

So we can really just shortcut this - we have two days, and we
can call `Seconds.secondsBetween` to get what we want.