Rename KickstarterClient to ApiClient
Retrofit clients have a set endpoint, e.g. https://api.kickstarter.com/
or https://www.kickstarter.com/. It's not designed to switch out endpoints
for different types of requests.

Our app will make a request to https://www.kickstarter.com/ to ping for build
updates, so we need a seperate class and client for that request.