Forwarding intent cleanup
Starting to like the idea of activities being responsible for launching new intents,
but presenters controlling *when* that happens.

e.g. an activity might respond to a click from a login button, it naively passes
the click onto the presenter. The presenter decides what should be done - if the
action involves starting a new activity, it should tell the original activity to start
up the new activity.

The presenter felt like the wrong place to launch new activities because every line
was reaching into the activity to do something - e.g. get a string for an intent, handle
a transition, launch the new activity.