Clean up spacing on filters
This is pretty tricky. To make things look perfect, we would need to set uneven margins
and paddings - but that causes problems with the touch area. The uneven margins and
paddings would also be revealed when holding down on the view for the ripple effect.

So instead, I've just tried to make it look acceptable but have nice behavior - clicks
should be associated with the correct view, ripples work well.