Disable missing translation lint
As soon as we add translations for another locale, this linter checks to make sure
all strings have translations in all locales. We want to be a little more flexible
than that - we can't guarantee that there will be a translation for each string,
and sometimes we don't even want a translation for strings (e.g: config strings)

It's possible to disable the lint check on a particular string using translatable="false"
but even then I don't think we want this just now.