Fix color mismatch on discovery cards and recycler view
The card layouts shouldn't need to set background colors, they can
just be transparent and let the background of the recycler view do
its thing.

This was causing an almost imperceptible change of color in margins
between cards, because the discovery background was slightly
different to the value that had been set on all the cards.