Add playground blank slate
When I experiment with new code, I find there's some mental overhead with writing it
in a new class; I need to think about which class to write it in, keep it separate from existing code,
remember to back it out. All small things, but each time it's a bit frustrating.

The playground code is intended as a scratchpad that removes those decisions; there's no logic in this
code so you're free to do what you like, and when you're finished just check out anything in
app/src/internal.