Fix AllowBackup lint warnings
"Fix". We don't have a lot of data to backup, just shared preferences. They
should be fine to restore. Was also receiving a warning about backing up GCM ids, but
after hunting around for 30 minutes I couldn't figure out how to take action on it.

The fact that I can simply add a link to a fullBackupContent file and the lint error
goes away tells me that there needs to be a more sophisticated check.