Add Reward card viewmodel
In preparation for reward itemization, I've added a ViewModel for the reward ViewHolder and rewritten the ViewHolder to be reactive. Otherwise the reward card should behave/look identical to the previous code.

Now that this code is tested, we can implement reward itemization with more confidence that existing functionality won't be broken. :heart: