Use setPrimaryItem to determine page
This should fix a bug where restoring the first discovery page after the
app has been stopped (not killed) results in a blank page.

The previous approach used `pagerCreatedPage` and `pageSelections`.
Unfortunately we can't rely entirely on `pageSelections` since it
will emit page 0 on subscribe, even if a different page is eventually
restored. `pagerCreatedPage` is also problematic because the fragment
manager may not *create* the fragment instance again on reload; the
fragment instance persists and we never get the create event.

It seems like the best tool available to figure out when a page is ready
to receive params is to hook into `setPrimaryItem` events. It fires many
times, so we use `distinctUntilChanged` to de-dupe. The one last wrinkle
is that it seems like a fragment may not be ready to receive params immediately
after `setPrimaryItem` is called – I've hooked up a delay so that execution
here is deferred until the next run loop. It's pretty janky and maybe we
could spend more time thinking about it since it's not a great pattern.