[🔥] Update pledge (#596)
* Added UpdateBacking mutation.
CurrentConfigType.getConfig was deprecated so I decided to officially retire it because it was causing crashes when it returned null. I am curious to see if it causes any lag.
While the updateBacking mutation is executing, a progress state is visible.
When a pledge is successfully updated, the ProjectActivity refreshes the project, pops the back stack and delegates the success to the BackingFragment. The BackingFragment shows a Snackbar with a success message.
When a pledge is unsuccessfully updated, the PledgeFragment shows a Snackbar with an error message.
The Update Pledge button in the PledgeFragment is only enabled when the user is updating their pledge and changed their shipping or pledge amount. It's always enabled when they are updating their reward.
Changed the root of fragment_backing to a CoordinatorLayout so the error Snackbar displays correctly.
Tests :)

Checking if the backing is processing vs the checkout state per CJ.
Fixed a bug where we didn't explicitly check for a valid pledge click before attempting to pledge or update pledge.