[♻️] NT-83 Project reload (#602)
* ProjectViewUtils.setActionButton does not show a project action button unless there are rewards.
The default project in ProjectFactory contains a list of rewards.
The rewards sheet is now visible in landscape.
The ProjectActivity now displays a progress bar when loading the project or refreshing. WOOHOO
ProjectIntentMapper.project no longer ignores errors.
Native checkout enabled projects can now be refreshed if the project failed to load.
OMG the tests. We need to test for deep links and coming from a parceled project.