[🏷] Adding helper methods for setting/getting ref tags from URLs (#620)
Renamed UrlUtils. buildUrl to UrlUtils.appendPath for clarity.
Added UrlUtils. appendQueryParameter as a helper method to add a query param to a URL.
Added UrlUtils. appendRefTag as a helper method to add a ref tag to a URL.
Added UrlUtils. refTag as a helper method to get a ref tag from a URL.
Tests