[🔁] NT-379 Synchronous createBacking (#637)
* CreateBacking mutation now determines a checkout with a successful state as a successful response.
I played myself by adding a delays to the Project page so I removed them. To properly fix a11y and not have any random delays, I just gave the pledge container root some initial margin so it's not at the same level of the toolbar and RecyclerView.
Instead of checking if the project is finishing and committing a FragmentTransaction, now just setting the BackingFragment and RewardsFragment as VISIBLE or GONE.
The pledge sheet can now be collapsed without animating which fixes having to delay showing the cancel pledge Snackbar. The expandPledgeSheet output in ProjectViewModel is now a Pair<Boolean, Boolean> where the first boolean is whether we're expanding and the second is whether we're animating.
Renamed ProjectActivity.animateRewards to expandPledgeSheet.
Added pledgeSuccessfullyCreated to the PledgeDelegate. ProjectViewModel now handles showing the ThanksActivity.
PledgeFragmentViewModel output startThanksActivity is now showPledgeSuccess that notifies the delegate.
Hitting the close button of the ThanksActivity when native checkout is enabled goes back to the ProjectActivity.
Tests.