[😭] NT-412 Pledge validation (#641)
* Added item_reward_pledge_card_disabled.xml which is basically identical to item_reward_pledge_card.xml but the Pledge button is not enabled. Calling RecyclerView.Adapter.notifyItemChanged does not guarantee that onBindViewHolder will be called again. But getItemViewType will so by returning a different layout, we can guarantee that the proper UI will be shown.
Fixed the bug with clicking the close card button in item_reward_pledge_card_disabled.xml and item_reward_pledge_card.xml
Moved pledge_minimum label under the Your pledge amount label and added pledge_maximum label that is initially INVISIBLE.
The maximum number of characters a user can enter in the pledge amount is 13. The maximum pledge (numerically) is 1,200,000.00 which is 12 characters including the separators but if the user didn't enter a decimal they could type 1,200,000,000.

Added output pledgeMaximum to PledgeFragmentViewModel that emits the maximum amount a user can pledge (including shipping).
Added output pledgeMaximumIsGone to PledgeFragmentViewModel that emits a Boolean determining if the maximum should be visible to the user. It returns true when the user has entered an amount that would put them over the project's country's currency maximum.
Added output pledgeButtonIsEnabled to PledgeFragmentViewModel that emits a Boolean determining if the Pledge button should be enabled. It returns true when the user has entered a valid pledge amount. It only emits for logged in users pledging or updating their payment method.
Added output continueButtonIsEnabled to PledgeFragmentViewModel that emits a Boolean determining if the Pledge button should be enabled. It returns true when the user is logged out and has entered a valid pledge amount. It only emits for logged out users.
Removed output showMinimumWarning from PledgeFragmentViewModel because we no longer need to warn them if they attempt to pledge an invalid amount; the button will be disabled.

* PR feedback from pairing with Nino:
Alpha.
continueButtonIsEnabled only emits for logged out users.
pledgeButtonIsEnabled only emits for logged in users who are pledging or updating their payment method.
Removing extra validation.
Cleanup for readability.

* fixed stepper tests and only emitting pledgeMax when not updating payment method