NT-457 A11y label for backing payment method (#647)
Added proper equals method to Either class. It compares the left and right values.
Added cardIssuer output to BackingFragmentViewModel. It's an Either<String, Int> representing the card brand or the string resource of the description.
Added Apple Pay content description string.
The payment method section in the BackingFragment is now focusable so it reads all at once.
Tests.