[👀] NT-510 Pledge CTA for creators viewing their own project is "View your rewards" (#665)
* Updated config and strings.
Renamed ProjectViewUtils.rewardsButtonColor to pledgeActionButtonColor. It now takes in a Project and a User.
Renamed ProjectViewUtils.rewardsButtonText to pledgeActionButtonText. It now takes in a Project and a User.
Renamed ProjectViewUtils.rewardsToolbarTitle to pledgeToolbarTitle. It now takes in a Project and a User.
Renamed ProjectViewModel output rewardsButtonColor to pledgeActionButtonColor.
Renamed ProjectViewModel output rewardsButtonText to pledgeActionButtonText .
Renamed project_action_button in the pledge container to pledge_action_button.
Renamed ProjectActivity.setProjectActionCTA to setPledgeActionButtonCTA.
Tests

* PR feedback from pairing with Justin: alpha and typo

* updated strings