Update layout root padding based on IME insets in `MessagesActivity`
Previously, when the device keyboard was opened it would cover the
messaging input field. Set the bottom padding of the layout's root view
based on IME insets to keep the input field visible.

This is the same logic as in `WindowInsetsUtil.manageEdgeToEdge()`
with a slight adjustment to handle the keyboard.