Manage Exceptions from Rx streams lacking `onError`
Any consumer of an Rx stream that does not handle errors creates an
opportunity for an unhandled `OnErrorNotImplementedException`.

Refactor the `RxJavaPlugins` error handler to instead report
all previously fatal Exceptions to Crashlytics. For `ApolloHttpException`s
with status code `429` in particular, report them with a custom key
to simplify querying.

The test case represents a scenario that is currently possible wherein a
`null` is "injected" into one of the app's Rx streams, leading to a fatal exception.