Add a 'Similar Projects Carousel' to ProjectOverviewFragment
To drive backer engagement, fetch and display a collection of
recommended projects on the main project page (non-prelaunch
projects).

Add `SimilarProjectsViewModel` to contain the logic for fetching and
maintaining state related to similar projects. This VM reacts to
whatever project is provided, but only fetches similar project data
once for a given project.

Add `SimilarProjectsComponent` to consume a collection of projects from
the ViewModel and display them in a Pager. This component leverages
custom touch event handling through a `pointerInput` Modifier to prevent
scroll gestures from being picked up by the parent `ViewPager2` of
`ProjectPageActivity`.